Remove HED 2 support (stage 1) and reorganize code
