--- common/schema/config.js
@@ -1,13 +1,0 @@
-/** Bundled HED schema configuration. */
-
-export const localSchemaList = new Map([
-  ['HED8.0.0', require('../../data/schemas/HED8.0.0.xml')],
-  ['HED8.1.0', require('../../data/schemas/HED8.1.0.xml')],
-  ['HED8.2.0', require('../../data/schemas/HED8.2.0.xml')],
-  ['HED8.3.0', require('../../data/schemas/HED8.3.0.xml')],
-  ['HED_score_1.0.0', require('../../data/schemas/HED_score_1.0.0.xml')],
-  ['HED_score_1.1.0', require('../../data/schemas/HED_score_1.1.0.xml')],
-  ['HED_score_2.0.0', require('../../data/schemas/HED_score_2.0.0.xml')],
-  ['HED_testlib_1.0.2', require('../../data/schemas/HED_testlib_1.0.2.xml')],
-  ['HED_testlib_2.0.0', require('../../data/schemas/HED_testlib_2.0.0.xml')],
-])
