+++ spec_tests/jsonTests.spec.js
@@ -4,8 +4,8 @@ import { beforeAll, describe, afterAll } from '@jest/globals'
 
 import * as hed from '../validator/event'
 import { BidsHedIssue } from '../bids/types/issues'
-import { buildSchemas } from '../validator/schema/init'
-import { SchemaSpec, SchemasSpec } from '../common/schema/types'
+import { buildSchemas } from '../schema/init'
+import { SchemaSpec, SchemasSpec } from '../schema/specs'
 import path from 'path'
 import { BidsSidecar, BidsTsvFile } from '../bids'
 import { generateIssue, IssueError } from '../common/issues/issues'
