+++ tests/utils/hed.spec.js
@@ -3,8 +3,8 @@ const assert = chai.assert
 import { beforeAll, describe, it } from '@jest/globals'
 
 import * as hed from '../../utils/hedStrings'
-import { SchemaSpec, SchemasSpec } from '../../common/schema/types'
-import { buildSchemas } from '../../validator/schema/init'
+import { SchemaSpec, SchemasSpec } from '../../schema/specs'
+import { buildSchemas } from '../../schema/init'
 
 describe('HED tag string utility functions', () => {
   describe('Syntactic utility functions', () => {
