+++ validator/event/index.js
@@ -1,23 +1,12 @@
 import { validateHedString, validateHedEvent, validateHedEventWithDefinitions } from './init'
 
-import { HedValidator } from './validator'
-import { Hed3Validator } from './hed3'
-import { Hed2Validator } from '../hed2/event/hed2Validator'
+import HedValidator from './validator'
 
-export {
-  validateHedString,
-  validateHedEvent,
-  validateHedEventWithDefinitions,
-  HedValidator,
-  Hed2Validator,
-  Hed3Validator,
-}
+export { validateHedString, validateHedEvent, validateHedEventWithDefinitions, HedValidator }
 
 export default {
   validateHedString,
   validateHedEvent,
   validateHedEventWithDefinitions,
   HedValidator,
-  Hed2Validator,
-  Hed3Validator,
 }
