+++ validator/index.js
@@ -1,7 +1,7 @@
 import { BidsDataset, BidsEventFile, BidsJsonFile, BidsSidecar, validateBidsDataset } from '../bids'
 import { validateHedDataset } from './dataset'
 import { validateHedEvent, validateHedString } from './event'
-import { buildSchemas } from './schema/init'
+import { buildSchemas } from '../schema/init'
 
 export {
   BidsDataset,
